requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.9.1
sentence-transformers==3.0.1
torch>=2.1.0
transformers>=4.41.0
accelerate>=0.33.0
FlagEmbedding==1.2.10
rank-bm25==0.2.2
tiktoken==0.7.0
lancedb==0.10.2
pyarrow>=16.1.0
pandas>=2.2.2
httpx==0.27.2
orjson==3.10.7
docling
numpy>=1.26.0
